requirements.txt
django
djangorestframework
boto3